fix: review feedback — security docs, multilingual intent, test split
Blocking:
- Add .tmp/ to .gitignore (prompt captures could be committed)
- Fix MILADY_SKIP_SECURITY_EVAL docs — env-var only, no dynamic
  channel detection. [SECURITY] prefix on startup warning.
- Add WARNING in JSDoc about multi-user deployments

Moderate:
- ISSUE_INTENT_RE: remove bare "close" and "label" (false positives)
- Verify buildEmptySocialExtractionResult schema matches upstream
- Add upstream prompt-prefix fragility comments

Multilingual:
- Add translated intent keywords for ko, zh-CN, es, pt, vi, tl
- 6 multilingual tests (Korean, Chinese, Spanish coding/terminal/issue)

Tests:
- Split security tests to prompt-security-heuristic.test.ts (90 LOC)
- Main test file now 454 LOC (was 508)
- 38 total tests across 2 files
